packages/pedagogy-engine/requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0

# API clients
anthropic>=0.18.0
openai>=1.0.0  # Used for OpenAI and DeepSeek (OpenAI-compatible API)

# LangChain v1 + LangGraph Agent Framework
langchain>=1.0.0
langgraph>=1.0.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Layer 4: Manim dataset validation
datasets>=2.14.0
manim>=0.18.0
docker>=6.0.0

#Hugging Face
sentence-transformers>=5.2.0

# Vector store
chromadb>=1.4.1
